FCL-151 | extract judgment header/put on styleguide
