+++ config/views/about.py
@@ -1,0 +1,13 @@
+from .template_view_with_context import TemplateViewWithContext
+from ds_caselaw_utils import courts
+
+
+class AboutThisServiceView(TemplateViewWithContext):
+    template_name = "pages/about_this_service.html"
+    page_title = "aboutthisservice.titleshort"
+
+    def get_context_data(self, **kwargs):
+        context = super().get_context_data(**kwargs)
+        context["courts"] = courts.get_listable_groups()
+        context["feedback_survey_type"] = "support"
+        return context
