+++ config/views/check.py
@@ -1,0 +1,5 @@
+from .template_view_with_context import TemplateViewWithContext
+
+
+class CheckView(TemplateViewWithContext):
+    template_name = "pages/check.html"
