+++ config/views/open_justice_license.py
@@ -1,0 +1,11 @@
+from .template_view_with_context import TemplateViewWithContext
+
+
+class OpenJusticeLicenceView(TemplateViewWithContext):
+    template_name = "pages/open_justice_licence.html"
+    page_title = "openjusticelicence.title"
+
+    def get_context_data(self, **kwargs):
+        context = super().get_context_data(**kwargs)
+        context["feedback_survey_type"] = "license"
+        return context
