+++ ds_judgements_public_ui/templates/includes/style_guide/judgment_header.html
@@ -1,0 +1,12 @@
+<h3 id="judgment-header">Judgment header</h3>
+<p>This is the static header that appears at the top of judgments. It has variants depending on the document_noun:</p>
+
+<p><strong>document_noun:</strong> "judgment"</p>
+<div class="style-guide__component-wrapper">
+  {% include "includes/judgment_text_toolbar.html" with document_noun="judgment" %}
+</div>
+
+<p><strong>document_noun:</strong> "press summary"</p>
+<div class="style-guide__component-wrapper">
+  {% include "includes/judgment_text_toolbar.html" with document_noun="press summary" %}
+</div>
