+++ ds_judgements_public_ui/templates/pages/style_guide.html
@@ -43,6 +43,7 @@
         {% include "includes/style_guide/detail_box.html" %}
         {% include "includes/style_guide/important_information_box.html" %}
         {% include "includes/style_guide/information_text.html" %}
+        {% include "includes/style_guide/judgment_header.html" %}
         {% include "includes/style_guide/notification_banners.html" %}
         {% include "includes/style_guide/radios.html" %}
         {% include "includes/style_guide/summary_card.html" %}
